requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
django==4.2.20
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.0
    # via -r requirements.in
faker==37.1.0
    # via -r requirements.in
numpy==2.2.5
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via faker
